Introduce a fallback encoding (eg. for blobs)
Usually you'll want to deliver the web pages using UTF-8. It's no
problem to convert Git's commit information to the PAGE_ENCODING since
if it's not UTF-8, the encoding that was used is specified.

In the case of blobs, Git by design doesn't want to know anything about
the encoding. But to make the file appear "normal" in the browser, the
FALLBACK_ENCODING (default: latin1) has a hint as to from which
encoding the string originates. In case the plain file is delivered, the
encoding will be set to the fallback specified if the blob's contents
are not valid UTF-8.

The same applies to the "Tagger" information, since tag objects don't
have an "encoding" field. (See:
http://git.661346.n2.nabble.com/PATCH-RFC-Document-format-of-basic-Git-objects-tp7287428p7288762.html )

Signed-off-by: Julius Plenz <plenz@cis.fu-berlin.de>